fastapi-art-protection/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pgvector
python-dotenv
pydantic
pydantic-settings
google-generativeai
alembic
numpy
Pillow
cryptography
python-multipart
